Demo App
Will send a HTTP Post request with methodName and two arguments. This will then be redirected back to the UIApplication delegate.

Using this pattern we can send instructions to the mock services. We will need to create a category on the appropriate UIApplication class. For HotPizza this would be "HPAppDelegate"